requirements.txt
absl-py==2.0.0
accelerate==0.24.1
aiohttp==3.9.0
aiosignal==1.3.1
albumentations==1.3.1
aniso8601==9.0.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autofaiss==2.15.8
beautifulsoup4==4.12.2
blinker==1.7.0
braceexpand==0.1.7
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.5.2
clip-retrieval==2.37.0
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
dataclasses==0.6
datasets==2.15.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
# Editable install with no version control (diffusers==0.24.0.dev0)
-e /workspace/zlatkd12/seulgi/Customization/src/diffusers
dill==0.3.7
dlib==19.24.2
docker-pycreds==0.4.0
embedding-reader==1.5.1
exceptiongroup==1.1.3
executing==2.0.1
ExifRead-nocycle==3.0.1
facenet-pytorch==2.5.3
faiss-cpu==1.7.4
filelock==3.13.1
fire==0.4.0
Flask==2.3.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.10
flatbuffers==2.0.7
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
gast==0.4.0
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
gunicorn==21.2.0
h5py==3.10.0
huggingface-hub==0.19.4
idna==3.4
imageio==2.33.0
img2dataset==1.42.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
intel-openmp==2024.0.0
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mediapipe==0.10.8
mkl==2024.0.0
mkl-service==2.4.0
ml-dtypes==0.2.0
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.0.4
multilingual-clip==1.0.10
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.15.0
open-clip-torch==2.23.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
prometheus-client==0.19.0
promise==2.3
prompt-toolkit==3.0.40
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==7.0.0
pyarrow-hotfix==0.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qudida==0.0.4
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.9.1
seaborn==0.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.35.0
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tbb==2021.11.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.11
tokenizers==0.13.3
torch==1.13.1
torchaudio==2.1.0
torchvision==0.14.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.28.0
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.216
urllib3==1.26.18
wandb==0.12.21
wcwidth==0.2.10
webdataset==0.2.77
Werkzeug==3.0.1
wrapt==1.14.1
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
